moved the package that the migration was in
