+++ src/main/resources/com/rackspace/salus/telemetry/dbmigrations/common/V1_0__initial.sql
