+++ src/main/resources/com/rackspace/salus/telemetry/dbmigrations/common/V1_3__add_translator_description.sql
