+++ src/main/resources/com/rackspace/salus/telemetry/dbmigrations/common/V4_0__index_monitor_label_selectors.sql
