+++ src/main/resources/com/rackspace/salus/telemetry/dbmigrations/common/V4_1__add_translator_order.sql
