+++ src/main/resources/com/rackspace/salus/telemetry/dbmigrations/h2/V1_2__align_baseline.sql
