+++ src/main/resources/com/rackspace/salus/telemetry/dbmigrations/h2/V3_0__change_monitor_type_in_translations.sql
