+++ src/main/resources/com/rackspace/salus/telemetry/dbmigrations/mysql/V1_1__hibernate_sequence.sql
