+++ src/main/resources/com/rackspace/salus/telemetry/dbmigrations/mysql/V1_2__align_baseline.sql
